+++ changelog.md
@@ -15,6 +15,7 @@ Fixes
 - Prevent creation of turns ending in generic crossings #511
 - Fix copycat headstock connectivity #512
 - Fix handcar/phantom track tooltips #658
+- Fix unfueled trains being unstoppable when realism is enabled #535
 ------------------------------------------------------
 Steam 'n' Rails 1.6.9
 ------------------------------------------------------